+++ chronosphere/intschema/slo.go
@@ -76,11 +76,12 @@ type SloSliCustomIndicator struct {
 
 type SloDefinition struct {
 	Objective              float64                               `intschema:"objective"`
-	ReportingWindows       []SloDefinitionReportingWindows       `intschema:"reporting_windows"`
 	BurnRateAlertingConfig []SloDefinitionBurnRateAlertingConfig `intschema:"burn_rate_alerting_config,optional,computed"`
+	EnableBurnRateAlerting bool                                  `intschema:"enable_burn_rate_alerting,optional,computed"`
+	TimeWindow             *SloDefinitionTimeWindow              `intschema:"time_window,optional,computed,list_encoded_object"`
 }
 
-type SloDefinitionReportingWindows struct {
+type SloDefinitionTimeWindow struct {
 	Duration string `intschema:"duration"`
 }